benchmarks/imagenet-e/ImageNet-Editing/requirements.txt
numpy
Pillow
opencv-python
torch
imageio
torchvision
einops
scikit-image
kornia
dominate
lpips
Ninja
ftfy
regex
matplotlib
imutils
gradio